cmd commit

+++ src/zadaci_21_08_2015/Zadatak5.java
@@ -29,7 +29,7 @@ public class Zadatak5 {
 			
 			// sabiranje bodova
 			while (inputFile.hasNext()) {
-				sum += inputFile.nextInt();
+				sum += inputFile.nextDouble();
 				counter++;
 			}
